refactor(access): Move manager to own dir
ref: #982 #983
